Switch from models.py to directory with multiple files
Refactor models.py to a directory for better organization,
ease of finding related models.

closes #5325
http://pulp.plan.io/issues/5325
