+++ pulp_rpm/app/models/__init__.py
@@ -1,0 +1,7 @@
+from .advisory import UpdateCollection, UpdateCollectionPackage, UpdateRecord, UpdateReference  # noqa
+from .comps import PackageCategory, PackageEnvironment, PackageGroup, PackageLangpacks  # noqa
+from .custom_metadata import RepoMetadataFile  # noqa
+from .distribution import Addon, Checksum, DistributionTree, Image, Variant  # noqa
+from .modulemd import Modulemd, ModulemdDefaults  # noqa
+from .package import Package  # noqa
+from .repository import RpmDistribution, RpmPublication, RpmRemote, RpmRepository  # noqa
